Fix landing page ingress to work for paths without trailing slash
Add an NginX snippet to the landing page ingress for path based deployments:
When the URL does not request a file and does not end with a `/`,
redirect to add a trailing `/`.
This allows correct resolution of relative paths to resources (e.g. js, css files).

Related to https://github.com/eclipsesource/theia-cloud/issues/124

Contributed on behalf of STMicroelectronics and CEA

Signed-off-by: Lucas Koehler <lkoehler@eclipsesource.com>