WIP: Add conversion webhook
Create deployment for java webhook.
Create certificate and pass the values to the webhook service.
Rename AppDefinition versions from vXbeta to v1betaX.

TODO:
The caBundle needs to be automatically consumed from the secret.
Check what needs to be parameterized in the values.
Adjust documentation and choose sensible default values.
Add webhook to other CRs as well.

Contributed on behalf of STMicroelectronics