Provide conversion webhook for CRDs
Webhook is called, whenever a CR is requested in a specific version.
This is enabled from these versions:
- `AppDefintion.v1beta8`
- `Session.v1beta6`
- `Workspace.v1beta3`
Older versions are deprecated and no longer in the definition.

Move status like fields to status:
- `Session.v1beta7`: Move `url`, `lastActivity` and `error` fields from
the spec to the status.
- `Workspace.v1beta4`: Move the `error` field from the spec to the status.
Also add the `error` field to `Workspace.v1beta3` as it was missing

Remove `timeout.strategy` from AppDefinition
- `AppDefinition.v1beta9`: Removed `timeout.strategy` and `timeout.limit`
is now just `timeout`. This was done, as there is only one Strategy left.

Contributed on behalf of STMicroelectronics
Co-authored-by: Johannes Faltermeier <jfaltermeier@eclipsesource.com>