Add configurable image preloading
Extends the main chart to facilitate preloading images on all nodes via a daemon set.
This increases startup performance on nodes which did not contain a session
because the application image will already be present.

All images configured in the values are preloaded.
One special case is handled: If no images are configured (the default)
but image preloading and the demo application are both enabled,
the demo image is still automatically preloaded.