Fix theia-cloud helm chart not being updateable
Properly quote the `rules.host` when loading from cluster on `helm upgrade`.
This is required, as a `*` is a special char in yaml and will break the parsing on `upgrade`.
`*` is a common wildcard character for host names so it should be supported.
With this the upgrading should work again.
Also add a note to the values file, that updating the `allWildcardInstances` will not work with `helm upgrade`, which is intended for now.