Allow service to read Deployments and ReplicaSets to trace owner refs
Read access to Deployments and ReplicaSets allows the service to trace owner references from Pods over ReplicaSets and Deployments to  the
Pod's Session.
This is required to find a Pod's corresponding Session (and vice-versa) in eager start mode. This is necessary because the Pod's ENV cannot contain the session name in eager start mode.